pyproject.toml
[tool.poetry]
name = "image-converter"
version = "0.1.0"
description = ""
authors = ["Bas <57452503+basekkelenkamp@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "image_converter"}]

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.2.0"
imageio = "^2.33.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
